Suicune done, darkness map

+++ src/battle_interface.c
@@ -911,7 +911,7 @@ void InitBattlerHealthboxCoords(u8 battler)
     UpdateSpritePos(gHealthboxSpriteIds[battler], x, y);
 }
 
-static void UpdateLvlInHealthbox(u8 healthboxSpriteId, u8 lvl)
+static void UpdateLvlInHealthbox(u8 healthboxSpriteId, u32 lvl)
 {
     u32 windowId, spriteTileNum;
     u8 *windowTileData;
